+++ src/CameraSensor.cc
@@ -351,8 +351,6 @@ bool CameraSensor::Update(const ignition::common::Time &_now)
     msg.set_pixel_format_type(msgsPixelFormat);
     msg.mutable_header()->mutable_stamp()->set_sec(_now.sec);
     msg.mutable_header()->mutable_stamp()->set_nsec(_now.nsec);
-
-
     auto frame = msg.mutable_header()->add_data();
     frame->set_key("frame_id");
     frame->add_value(this->Name());